refactor common code between TX and DB
