+++ js/browser/wa-crsqlite/src/DB.ts
@@ -1,9 +1,8 @@
-import * as SQLite from "@vlcn.io/wa-sqlite";
-import { DBAsync, StmtAsync, UpdateType } from "@vlcn.io/xplat-api";
+import { DBAsync, StmtAsync, TXAsync, UpdateType } from "@vlcn.io/xplat-api";
 import { SQLITE_UTF8 } from "@vlcn.io/wa-sqlite";
-import { serialize, serializeTx, topLevelMutex } from "./serialize.js";
+import { serialize, topLevelMutex } from "./serialize.js";
 import Stmt from "./Stmt.js";
-import { computeCacheKey } from "./cache.js";
+import TX from "./TX.js";
 
 export class DB implements DBAsync {
   public readonly __mutex = topLevelMutex;
@@ -26,8 +25,17 @@ export class DB implements DBAsync {
     (type: UpdateType, dbName: string, tblName: string, rowid: bigint) => void
   > | null = null;
   #closed = false;
+  #tx: TX;
 
-  constructor(public api: SQLiteAPI, public db: number) {}
+  constructor(public api: SQLiteAPI, public db: number) {
+    this.#tx = new TX(
+      api,
+      db,
+      topLevelMutex,
+      this.#assertOpen,
+      this.stmtFinalizer
+    );
+  }
 
   get siteid(): string {
     return this.#siteid!;
@@ -41,69 +49,17 @@ export class DB implements DBAsync {
   }
 
   execMany(sql: string[]): Promise<any> {
-    return serialize(
-      this.cache,
-      null,
-      () => this.api.exec(this.db, sql.join("")),
-      this.__mutex
-    );
+    return this.#tx.execMany(sql);
   }
 
   exec(sql: string, bind?: SQLiteCompatibleType[]): Promise<void> {
-    // TODO: either? since not returning?
-    this.#assertOpen();
-    return serialize(
-      this.cache,
-      computeCacheKey(sql, "a", bind),
-      () => {
-        return this.statements(sql, false, bind);
-      },
-      this.__mutex
-    );
+    return this.#tx.exec(sql, bind);
   }
 
-  #assertOpen() {
+  #assertOpen = () => {
     if (this.#closed) {
       throw new Error("The DB is closed");
     }
-  }
-
-  onUpdate(
-    cb: (
-      type: UpdateType,
-      dbName: string,
-      tblName: string,
-      rowid: bigint
-    ) => void
-  ): () => void {
-    if (this.#updateHooks == null) {
-      this.api.update_hook(this.db, this.#onUpdate);
-      this.#updateHooks = new Set();
-    }
-    this.#updateHooks.add(cb);
-
-    return () => this.#updateHooks?.delete(cb);
-  }
-
-  #onUpdate = (
-    type: UpdateType,
-    dbName: string,
-    tblName: string,
-    rowid: bigint
-  ) => {
-    if (this.#updateHooks == null) {
-      return;
-    }
-    this.#updateHooks.forEach((h) => {
-      // we wrap these since listeners can be thought of as separate threads of execution
-      // one dieing shouldn't prevent others from being notified.
-      try {
-        h(type, dbName, tblName, rowid);
-      } catch (e) {
-        console.error("Failed notifying a DB update listener");
-        console.error(e);
-      }
-    });
   };
 
   /**
@@ -113,13 +69,7 @@ export class DB implements DBAsync {
     sql: string,
     bind?: SQLiteCompatibleType[]
   ): Promise<T[]> {
-    this.#assertOpen();
-    return serialize(
-      this.cache,
-      computeCacheKey(sql, "o", bind),
-      () => this.statements(sql, true, bind),
-      this.__mutex
-    );
+    return this.#tx.execO(sql, bind);
   }
 
   /**
@@ -129,44 +79,15 @@ export class DB implements DBAsync {
     sql: string,
     bind?: SQLiteCompatibleType[]
   ): Promise<T[]> {
-    this.#assertOpen();
-    return serialize(
-      this.cache,
-      computeCacheKey(sql, "a", bind),
-      () => this.statements(sql, false, bind),
-      this.__mutex
-    );
+    return this.#tx.execA(sql, bind);
   }
 
   prepare(sql: string): Promise<StmtAsync> {
-    this.#assertOpen();
-    return serialize(
-      this.cache,
-      undefined,
-      async () => {
-        const str = this.api.str_new(this.db, sql);
-        const prepared = await this.api.prepare_v2(
-          this.db,
-          this.api.str_value(str)
-        );
-        if (prepared == null) {
-          this.api.str_finish(str);
-          throw new Error(`Could not prepare ${sql}`);
-        }
+    return this.#tx.prepare(sql);
+  }
 
-        return new Stmt(
-          this,
-          this.stmtFinalizer,
-          // this.stmtFinalizationRegistry,
-          this.cache,
-          this.api,
-          prepared.stmt,
-          str,
-          sql
-        );
-      },
-      this.__mutex
-    );
+  transaction(cb: (tx: TXAsync) => Promise<void>): Promise<void> {
+    return this.#tx.transaction(cb);
   }
 
   /**
@@ -212,84 +133,41 @@ export class DB implements DBAsync {
     );
   }
 
-  async savepoint(cb: () => Promise<void>): Promise<void> {
-    throw new Error("do not use me yet");
-    this.#assertOpen();
-    await this.exec("SAVEPOINT");
-    await cb();
+  onUpdate(
+    cb: (
+      type: UpdateType,
+      dbName: string,
+      tblName: string,
+      rowid: bigint
+    ) => void
+  ): () => void {
+    if (this.#updateHooks == null) {
+      this.api.update_hook(this.db, this.#onUpdate);
+      this.#updateHooks = new Set();
+    }
+    this.#updateHooks.add(cb);
+
+    return () => this.#updateHooks?.delete(cb);
   }
 
-  transaction(cb: (tx: DBAsync) => Promise<void>): Promise<void> {
-    this.#assertOpen();
-    return serializeTx(
-      async (tx: DBAsync) => {
-        await this.exec("SAVEPOINT crsql_transaction");
-        try {
-          await cb(tx);
-        } catch (e) {
-          await this.exec("ROLLBACK");
-          return;
-        }
-        await this.exec("RELEASE crsql_transaction");
-      },
-      this.__mutex,
-      this
-    );
-  }
-
-  private async statements(
-    sql: string,
-    retObjects: boolean,
-    bind?: unknown[]
-  ): Promise<any> {
-    const results: { columns: string[]; rows: any[] }[] = [];
-
-    for await (const stmt of this.api.statements(this.db, sql)) {
-      const rows: any[] = [];
-      const columns = this.api.column_names(stmt);
-      if (bind) {
-        this.bind(stmt, bind);
-      }
-      while ((await this.api.step(stmt)) === SQLite.SQLITE_ROW) {
-        const row = this.api.row(stmt);
-        rows.push(row);
-      }
-      if (columns.length) {
-        results.push({ columns, rows });
-      }
+  #onUpdate = (
+    type: UpdateType,
+    dbName: string,
+    tblName: string,
+    rowid: bigint
+  ) => {
+    if (this.#updateHooks == null) {
+      return;
     }
-
-    // we'll only return results for first stmt
-    // if (results.length > 1) {
-    //   throw new Error("We currently only support 1 statement per query.");
-    // }
-    const returning = results[0];
-    if (returning == null) return null;
-
-    if (!retObjects) {
-      return returning.rows;
-    }
-
-    const objects: Object[] = [];
-    for (const row of returning.rows) {
-      const o: { [key: string]: any } = {};
-      for (let i = 0; i < returning.columns.length; ++i) {
-        o[returning.columns[i]] = row[i];
+    this.#updateHooks.forEach((h) => {
+      // we wrap these since listeners can be thought of as separate threads of execution
+      // one dieing shouldn't prevent others from being notified.
+      try {
+        h(type, dbName, tblName, rowid);
+      } catch (e) {
+        console.error("Failed notifying a DB update listener");
+        console.error(e);
       }
-      objects.push(o);
-    }
-
-    return objects;
-  }
-
-  private bind(stmt: number, values: unknown[]) {
-    for (let i = 0; i < values.length; ++i) {
-      const v = values[i];
-      this.api.bind(
-        stmt,
-        i + 1,
-        typeof v === "boolean" ? (v && 1) || 0 : (v as any)
-      );
-    }
-  }
+    });
+  };
 }
